Add support for systemd socket activation
Adding systemd support allows spawning polipo in parallel with the services
that depend on polipo.

The caller (spawner) owns the sockets, then spawns polipo and tells polipo
which sockets it has to use. (Before this change, you'd have to spawn polipo,
ask it which sockets it's using, and only then spawn the services that depend
on polipo.)

sd-daemon.[ch] taken verbatim from systemd as recommended in "systemd for
Developers"[1]

[1] http://0pointer.de/blog/projects/socket-activation.html